EY-1768: AzureAdClient-fila byrja å bli lang og omfattande, så trekkjer ut dataklassane for tokens til ei eiga fil
